requirements.txt
absl-py==1.4.0
accelerate==0.16.0
aiohttp==3.9.5
aiosignal==1.2.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.2.0
audioread==3.0.0
autogluon==0.7.0
autogluon.common==0.7.0
autogluon.core==0.7.0
autogluon.features==0.7.0
autogluon.multimodal==0.7.0
autogluon.tabular==0.7.0
autogluon.timeseries==0.7.0
backcall==0.2.0
beautifulsoup4==4.11.2
bert-score==0.3.13
bitarray==2.7.1
blessed==1.20.0
blinker==1.6.2
blis==0.7.9
boto3==1.26.81
botocore==1.29.81
Brotli==1.0.9
cachetools==5.3.0
catalogue==2.0.8
catboost==1.1.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
crcmod==1.7
cryptography==42.0.5
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
dataclasses==0.6
datasets==2.10.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
editdistance==0.6.2
evaluate==0.3.0
executing==0.8.3
fairscale==0.4.13
fairseq==0.10.0
fastai==2.7.11
fastcore==1.5.28
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.14.0
fonttools==4.38.0
frozenlist==1.4.0
fsspec==2023.1.0
future==0.18.3
gdown==4.6.4
gluonts==0.12.3
google-auth==2.16.0
google-auth-oauthlib==0.5.2
gpustat==1.0.0
graphviz==0.20.1
grpcio==1.48.2
huggingface-hub==0.12.0
hydra-core==1.0.7
hyperopt==0.2.7
idna==3.7
imageio==2.26.0
importlib-metadata==7.0.1
importlib-resources==5.10.2
ipykernel==6.15.0
ipython==8.12.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.7.2
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.2
lightgbm==3.3.5
lightning-utilities==0.7.1
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
model-index==0.1.11
msgpack==1.0.4
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nest-asyncio==1.6.0
networkx==2.8.8
nlpaug==1.1.11
nltk==3.8.1
nptyping==2.4.1
numba==0.56.4
numpy==1.23.5
nvidia-ml-py==11.495.46
oauthlib==3.2.2
omegaconf==2.2.3
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.13.1
pmdarima==1.8.5
pooch==1.6.0
portalocker==2.7.0
preshed==3.0.8
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.5
pyDeprecate==0.3.2
Pygments==2.14.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.3
pytorch-metric-learning==1.7.3
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.2
ray==2.2.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.0
resampy==0.4.2
responses==0.18.0
rich==13.4.2
rsa==4.7.2
s3transfer==0.6.0
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
sentencepiece==0.1.96
seqeval==1.2.2
setuptools==60.2.0
six==1.16.0
sktime==0.15.1
smart-open==6.3.0
soundfile==0.11.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.2.0
statsforecast==1.4.0
statsmodels==0.13.5
tabulate==0.9.0
tbats==1.1.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
text-unidecode==1.3
thinc==8.1.7
threadpoolctl==2.2.0
tifffile==2023.2.28
timm==0.6.12
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1+cu116
torchmetrics==0.8.2
torchvision==0.14.1+cu116
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.26.1
typer==0.7.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.7.0
urllib3==1.26.19
virtualenv==20.20.0
wasabi==1.1.1
wcwidth==0.2.5
Werkzeug==2.3.4
wheel==0.38.4
wrapt==1.15.0
xgboost==1.7.4
xxhash==3.2.0
yarl==1.8.2
zipp==3.12.1